move console role to private header
increase value of last console role, to allow for more roles in future
